IA_devs/AI_applying_Python/requirements.txt
google-genai
python-dotenv
openai
requests